+++ app/Dto/EmployeeManagement/Employee/EmployeeShowDto.php
@@ -1,19 +1,23 @@
 <?php
 
-namespace App\Dto\Employee;
+namespace App\Dto\EmployeeManagement\Employee;
 
 use App\Dto\CoreDto;
 
-class EmployeeUpdateDto extends CoreDto
+class EmployeeShowDto extends CoreDto
 {
     public int $id;
     public string $full_name;
 
     public int $type_document_id;
 
+    public string $type_document_name;
+
     public string $document_number;  
 
     public int $municipality_id;
+
+    public string $municipality_name;
 
     public string $address;
 
@@ -21,13 +25,23 @@ class EmployeeUpdateDto extends CoreDto
 
     public int $gender_id;
 
-    public int $job_position_id;
+    public string $gender_name;
+
+    public ?int $job_position_id;
+
+    public ?string $job_position_name;
 
     public int $eps_id;
 
+    public string $eps_name;
+
     public int $pension_fund_id;
 
+    public string $pension_fund_name;
+
     public int $arl_id;
+
+    public string $arl_name;
 
     public int $contract_type_id;
 
@@ -39,13 +53,21 @@ class EmployeeUpdateDto extends CoreDto
 
     public int $bank_id;
 
+    public string $bank_name;
+
     public ?string $bank_account_number;
 
     public string $emergency_contact;
 
-    public int $employee_status_id;
+    public ?int $user_who_created_id ;
+
+    public ?string $user_who_created_name;
 
     public ?int $user_who_updated_id;
+
+    public ?string $user_who_updated_name;
+
+    public ?string $created_at;
 
     public ?string $updated_at;
 }
